+++ src/pages/mypage/hooks/useMyPagePasswordChange.ts
@@ -1,33 +1,23 @@
 import type { FormEvent } from 'react';
 import { useEffect, useMemo, useState } from 'react';
-import { useNavigate } from 'react-router';
 
-import { ROUTES } from '../../../constants/routes';
 import {
   extractAuthApiErrorMessage,
   extractAuthApiFieldErrors,
 } from '../../../features/auth/api/auth';
-import {
-  useChangePasswordMutation,
-  useDeleteAccountMutation,
-} from '../../../features/auth/api/useAuthApi';
-import { clearAuthSession } from '../../../features/auth/utils/sessionManager';
+import { useChangePasswordMutation } from '../../../features/auth/api/useAuthApi';
 import type {
   PasswordChangeFieldName,
   PasswordChangeValues,
 } from '../../../components/mypage/PasswordChangePanel';
-import type { MyPageToastPayload } from '../types';
 
 type PasswordTouchedState = Record<PasswordChangeFieldName, boolean>;
 type PasswordFieldErrors = Partial<Record<PasswordChangeFieldName, string>>;
+
 type PasswordPanelMessage = {
   tone: 'success' | 'error';
   message: string;
 } | null;
-
-type UseMyPageSecurityOptions = {
-  onToast: (toast: MyPageToastPayload) => void;
-};
 
 const initialPasswordValues: PasswordChangeValues = {
   currentPassword: '',
@@ -82,10 +72,8 @@ const mapPasswordApiFieldErrors = (
   newPasswordConfirm: fieldErrors.new_password_check,
 });
 
-function useMyPageSecurity({ onToast }: UseMyPageSecurityOptions) {
-  const navigate = useNavigate();
+function useMyPagePasswordChange() {
   const changePasswordMutation = useChangePasswordMutation();
-  const deleteAccountMutation = useDeleteAccountMutation();
   const [isPasswordPanelOpen, setIsPasswordPanelOpen] = useState(false);
   const [passwordValues, setPasswordValues] = useState<PasswordChangeValues>(
     initialPasswordValues,
@@ -95,9 +83,6 @@ function useMyPageSecurity({ onToast }: UseMyPageSecurityOptions) {
   const [apiFieldErrors, setApiFieldErrors] = useState<PasswordFieldErrors>({});
   const [passwordPanelMessage, setPasswordPanelMessage] =
     useState<PasswordPanelMessage>(null);
-  const [isDeleteModalOpen, setIsDeleteModalOpen] = useState(false);
-  const [deletePassword, setDeletePassword] = useState('');
-  const [deletePasswordError, setDeletePasswordError] = useState('');
   const localFieldErrors = useMemo(
     () => getPasswordFieldErrors(passwordValues, touchedState),
     [passwordValues, touchedState],
@@ -235,68 +220,6 @@ function useMyPageSecurity({ onToast }: UseMyPageSecurityOptions) {
     }
   };
 
-  const openDeleteModal = () => {
-    setDeletePassword('');
-    setDeletePasswordError('');
-    setIsDeleteModalOpen(true);
-  };
-
-  const closeDeleteModal = () => {
-    setIsDeleteModalOpen(false);
-    setDeletePassword('');
-    setDeletePasswordError('');
-  };
-
-  const handleDeletePasswordChange = (value: string) => {
-    setDeletePassword(value);
-
-    if (deletePasswordError) {
-      setDeletePasswordError('');
-    }
-  };
-
-  const handleDeleteAccount = async () => {
-    const trimmedPassword = deletePassword.trim();
-
-    if (!trimmedPassword) {
-      setDeletePasswordError('현재 비밀번호를 입력해주세요.');
-      return;
-    }
-
-    try {
-      await deleteAccountMutation.mutateAsync({
-        password: trimmedPassword,
-      });
-
-      clearAuthSession();
-      navigate(`/${ROUTES.LOGIN}`, {
-        replace: true,
-        state: {
-          noticeMessage: '회원 탈퇴가 완료되었습니다.',
-        },
-      });
-    } catch (error) {
-      const fieldErrors = extractAuthApiFieldErrors(error);
-
-      if (fieldErrors.password) {
-        setDeletePasswordError(fieldErrors.password);
-        return;
-      }
-
-      const errorMessage = extractAuthApiErrorMessage(error);
-
-      if (errorMessage.includes('비밀번호')) {
-        setDeletePasswordError(errorMessage);
-        return;
-      }
-
-      onToast({
-        tone: 'error',
-        message: errorMessage,
-      });
-    }
-  };
-
   return {
     isPasswordPanelOpen,
     togglePasswordPanel,
@@ -308,15 +231,7 @@ function useMyPageSecurity({ onToast }: UseMyPageSecurityOptions) {
     handlePasswordBlur,
     handlePasswordSubmit,
     isChangePasswordPending: changePasswordMutation.isPending,
-    isDeleteModalOpen,
-    openDeleteModal,
-    closeDeleteModal,
-    deletePassword,
-    deletePasswordError,
-    handleDeletePasswordChange,
-    handleDeleteAccount,
-    isDeleteAccountPending: deleteAccountMutation.isPending,
   };
 }
 
-export default useMyPageSecurity;
+export default useMyPagePasswordChange;
